requirements.txt
Pillow==7.2.0
numpy